Move function to a more approptiate file
