+++ activities/audio.go
@@ -87,40 +87,6 @@ func DetectSilence(ctx context.Context, input common.AudioInput) (bool, error) {
 	return transcode.AudioIsSilent(input.Path)
 }
 
-type GetVideoOffsetInput struct {
-	VideoPath1 paths.Path
-	VideoPath2 paths.Path
-}
-
-func GetVideoOffset(ctx context.Context, input GetVideoOffsetInput) (int, error) {
-	log := activity.GetLogger(ctx)
-	activity.RecordHeartbeat(ctx, "GetVideoOffset")
-	log.Info("Starting GetVideoOffsetActivity")
-
-	video1TC, err := ffmpeg.GetTimeCode(input.VideoPath1.Local())
-	if err != nil {
-		return 0, err
-	}
-
-	video2TC, err := ffmpeg.GetTimeCode(input.VideoPath2.Local())
-	if err != nil {
-		return 0, err
-	}
-
-	// Video from YouPlay is always 25fps and 48000Hz
-	videoSamples1, err := utils.TCToSamples(video1TC, 25, 48000)
-	if err != nil {
-		return 0, err
-	}
-
-	videoSamples2, err := utils.TCToSamples(video2TC, 25, 48000)
-	if err != nil {
-		return 0, err
-	}
-
-	return videoSamples2 - videoSamples1, nil
-}
-
 type ExtractAudioInput struct {
 	VideoPath       paths.Path
 	OutputFolder    paths.Path
